+++ config/splat.pspeu.dra.yaml
@@ -103,8 +103,6 @@ segments:
       - [0x35060, c, dra_psp/35060]
       - [0x353B0, c, dra/704D0] # O0
       - [0x373F8, c, dra_psp/373F8]
-      - [0x375C0, c, dra_psp/375C0]
-      - [0x37930, c, dra_psp/37930]
       - [0x39AA8, c, dra_psp/39AA8]
       - [0x3A4C0, c, dra_psp/3A4C0]
       - [0x3A770, c, dra_psp/3A770]
