+++ config/symbols.pspeu.dra.txt
@@ -128,6 +128,7 @@ CheckTetraSpiritInput = 0x09112E48;
 CheckSoulStealInput = 0x09113268;
 CheckSwordBrothersInput = 0x091136F8;
 CheckSpecialMoveInputs = 0x09113C00;
+func_8010E570 = 0x09113F40;
 func_8010E83C = 0x091142B0;
 func_8010E940 = 0x091143D0;
 DoGravityJump = 0x09114438;
