+++ src/dra_psp/373F8.c
@@ -8,3 +8,45 @@ INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", func_psp_09113D78);
 INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", func_psp_09113DE0);
 
 INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", func_psp_09113E70);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", func_8010E570);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", func_psp_091140D8);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", func_psp_09114208);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", func_8010E83C);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", func_8010E940);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", DoGravityJump);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", func_psp_09114500);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", func_psp_09114590);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", func_psp_09114668);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", func_psp_09114880);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", func_psp_091149C8);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", func_psp_09114A38);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", func_8010FAF4);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", PerformHellfire);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", PerformDarkMetamorphosis);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", PerformSoulSteal);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", PerformSummonSpirit);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", PerformTetraSpirit);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", PerformSwordBrothers);
+
+INCLUDE_ASM("dra_psp/psp/dra_psp/373F8", func_8010FD88);
+
+// Next func in PSP is SetPlayerStep; next in US is func_8010FDF8. File split.
