--- src/dra_psp/375C0.c
@@ -1,8 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../dra/dra.h"
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/375C0", func_psp_09113F40);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/375C0", func_psp_091140D8);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/375C0", func_psp_09114208);
