--- src/dra_psp/37930.c
@@ -1,38 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../dra/dra.h"
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/37930", func_8010E83C);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/37A50", func_8010E940);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/37A50", DoGravityJump);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/37A50", func_psp_09114500);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/37A50", func_psp_09114590);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/37A50", func_psp_09114668);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/37A50", func_psp_09114880);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/37A50", func_psp_091149C8);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/37A50", func_psp_09114A38);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/37A50", func_8010FAF4);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/37A50", PerformHellfire);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/37A50", PerformDarkMetamorphosis);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/37A50", PerformSoulSteal);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/37A50", PerformSummonSpirit);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/37A50", PerformTetraSpirit);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/37A50", PerformSwordBrothers);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/37A50", func_8010FD88);
-
-// Next func in PSP is SetPlayerStep; next in US is func_8010FDF8. File split.
